refactor: core/index.ts 추가 및 Import 구문에 적용
