--- src/adapter/todoAdapter.js
@@ -1,3 +1,0 @@
-import { RestClient } from "../core/RestClient";
-
-export const todoAdapterClient = new RestClient('https://blackcoffee-todolist.df.r.appspot.com/api');
