+++ src/components/Team/TeamAppendForm.js
@@ -1,4 +1,4 @@
-import {Component} from "../../core/Component";
+import {Component} from "../../core";
 import {ADD_TEAM, SET_OPENED_APPEND_FORM, teamStore} from "../../store/teamStore";
 
 export const TeamAppendForm = class extends Component {
