+++ src/components/Todo/TodoHeader.js
@@ -1,4 +1,4 @@
-import {Component} from "../../core/Component";
+import {Component} from "../../core";
 import {todoOfTeamStore} from "../../store/todoOfTeamStore";
 
 export const TodoHeader = class extends Component {
