+++ src/components/Todo/TodoItemAppender.js
@@ -1,4 +1,4 @@
-import {Component} from "../../core/Component";
+import {Component} from "../../core";
 import {ADD_ITEM, todoOfTeamStore} from "../../store/todoOfTeamStore";
 
 export const TodoItemAppender = class extends Component {
