+++ src/components/Todo/TodoList.js
@@ -1,9 +1,8 @@
-import {Component} from "../../core/Component";
+import {Component} from "../../core";
 import {DELETE_ITEM, SET_EDITING, TOGGLE_ITEM, UPDATE_ITEM, UPDATE_ITEM_PRIORITY, todoOfTeamStore} from "../../store/todoOfTeamStore";
 import {TodoListFooter} from "./TodoListFooter";
 import {TodoItemAppender} from "./TodoItemAppender";
-import PriorityTypes from "../../constants/PriorityTypes";
-import FilterTypes from "../../constants/FilterTypes";
+import { PriorityTypes, FilterTypes } from "../../constants";
 
 const priorityChip = {
   [PriorityTypes.PRIMARY]: 'primary',
