+++ src/components/Todo/TodoListOfTeam.js
@@ -1,4 +1,4 @@
-import {Component} from "../../core/Component";
+import {Component} from "../../core";
 import {TodoMemberAppender} from "./TodoMemberAppender";
 import {TodoList} from "./TodoList";
 import {todoOfTeamStore} from "../../store/todoOfTeamStore";
