+++ src/core/index.ts
@@ -1,0 +1,4 @@
+export * from './Component';
+export * from './RestClient';
+export * from './Router';
+export * from './Store';
